requirements.txt
Flask==2.0.2
pydub==0.25.1
SpeechRecognition==3.8.1
pytesseract==0.3.8
Pillow==8.4.0
PyPDF2==1.26.0
reportlab==3.6.4
transformers==4.15.0
torch==1.10.1
